requirements.txt
scikit-learn
pandas
matplotlib
tensorflow
spacy
streamlit
Pillow 